+++ src/renderer/components/MessageItem.vue
@@ -68,13 +68,12 @@
             <button @click="cancelEditing" class="tertiary">{{ t('chat.cancel') }}</button>
           </div>
         </div>
-        <MessageItemBody v-else :message="message" :show-tool-calls="showToolCalls" @media-loaded="onMediaLoaded" />
+        <MessageItemBody v-else :message="message" :tool-calls-display="toolCallsDisplay" @media-loaded="onMediaLoaded" />
       </div>
 
       <!-- transient information -->
-      <div v-if="message.transient" class="message-transient">
-        <MessageItemToolBlock v-for="runningTool in runningTools" :tool-call="runningTool" v-if="runningTools" />
-        <Loader v-if="!message.status" />
+      <div v-if="message.transient && !message.status" class="message-transient">
+        <Loader />
       </div>
 
     </div>
@@ -93,7 +92,6 @@ import Loader from '@components/Loader.vue'
 import MessageItemActions from '@components/MessageItemActions.vue'
 import MessageItemBody from '@components/MessageItemBody.vue'
 import MessageItemMediaBlock from '@components/MessageItemMediaBlock.vue'
-import MessageItemToolBlock from '@components/MessageItemToolBlock.vue'
 import SpinningIcon from '@components/SpinningIcon.vue'
 import useEventBus from '@composables/event_bus'
 import Attachment from '@models/attachment'
@@ -104,7 +102,7 @@ import type { ChatCallbacks } from '@screens/Chat.vue'
 import { t } from '@services/i18n'
 import { store } from '@services/store'
 import { LoaderCircleIcon } from 'lucide-vue-next'
-import { ChatToolMode } from 'types/config'
+import { ToolCallsDisplay } from 'types/config'
 import { computed, inject, onBeforeUnmount, onMounted, PropType, ref, watch } from 'vue'
 // import { getMarkdownSelection } from '@services/markdown'
 
@@ -137,7 +135,7 @@ const emits = defineEmits(['media-loaded'])
 const div = ref<HTMLElement|null>(null)
 const hovered = ref(false)
 const audio = ref<HTMLAudioElement|null>(null)
-const showToolCalls = ref<ChatToolMode>('always')
+const toolCallsDisplay = ref<ToolCallsDisplay>('summary')
 const audioState = ref<{state: string, messageId: string|null}>({
   state: 'idle',
   messageId: null,
@@ -152,9 +150,9 @@ onMounted(() => {
 
   // show tool calls
   if (props.showToolCalls) {
-    showToolCalls.value = store.config.appearance.chat.showToolCalls
+    toolCallsDisplay.value = store.config.appearance.chat.toolCallsDisplay
   } else {
-    showToolCalls.value = 'never'
+    toolCallsDisplay.value = 'none'
   }
 
   // make sure links are going outside
@@ -173,8 +171,8 @@ onMounted(() => {
   onBusEvent('audio-noise-detected', () =>  audioPlayer.stop)
 
   // settings change
-  watch(() => store.config.appearance.chat.showToolCalls, (value) => {
-    showToolCalls.value = value
+  watch(() => store.config.appearance.chat.toolCallsDisplay, (value) => {
+    toolCallsDisplay.value = value
   })
 
 })
@@ -233,11 +231,6 @@ const imageUrl = computed(() => {
 
 })
 
-const runningTools = computed(() => {
-  if (store.config.appearance.chat.showToolCalls === 'never') return null
-  const runningTools = props.message.toolCalls.filter(toolCall => !toolCall.done)
-  return runningTools.length ? runningTools : null
-})
 
 // using simple css :hover
 // was not working from a testing perspective
@@ -289,10 +282,10 @@ const readAloudText = async (text: string) => {
 }
 
 const onShowTools = () => {
-  if (showToolCalls.value !== 'always') {
-    showToolCalls.value = 'always'
+  if (toolCallsDisplay.value !== 'details') {
+    toolCallsDisplay.value = 'details'
   } else {
-    showToolCalls.value = 'never'
+    toolCallsDisplay.value = 'none'
   }
 }